refactor: Move file App for folder src.
