+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import Routes from './src/routes';
+import Routes from './routes/index';
 import {NavigationContainer} from '@react-navigation/native';
-import {AuthProvider} from './src/contexts/AuthContext';
+import {AuthProvider} from './contexts/AuthContext';
 
 const App = () => {
   return (
